requirements.txt
pyTelegramBotAPI
torch==1.10.1+cu102
torchvision==0.11.2+cu102
torchaudio===0.10.1+cu102
numpy
Pillow
scikit-image
fastai
